pyproject.toml

[project]
name = "steblikcom-django"
version = "0.1.0"
description = "steblik.com Django project"
readme = "README.md"
authors = [{ name = "miroslawsteblik", email = "miroslawsteblik@gmail.com" }]
requires-python = ">=3.12"

dependencies = [
    "django>=5.1,<6.0",
    "django-environ>=0.11",
    "whitenoise>=6.7",
    "markdown2>=2.5",
    "python-frontmatter>=1.1",
    "pygments>=2.18",
    "bleach>=6.0",
    "pillow>=12.2.0",
    "django-allauth>=65.16.1",
    "markdown>=3.10.2",
    "gunicorn>=23.0",
    "resend>=2.29.0",
    "django-extensions>=4.1",
    "django-htmx>=1.27.0",
    "django-csp>=4.0",
    "argon2-cffi>=25.1.0",
    "psycopg[binary]>=3.2",
]

[dependency-groups]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "django-debug-toolbar>=4.4",
    "pytest-django>=4.9",

    # Linting / formatting
    "ruff>=0.9.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
    "djlint>=1.36.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]


# ── pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "--strict-markers -q"
pythonpath = ["apps/web"]
testpaths = ["apps/web"]

# ── bandit ───────────────────────────────────────────────────────────────────
[tool.bandit]
exclude_dirs = [".venv", "tests"]
skips = ["B101"]                  # assert usage is fine in tests


# ── ruff ─────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "ANN", # flake8-annotations (type hints)
]
# ANN002/003 (*args/**kwargs) and ANN204 (special methods) are noise; everything else enforced
ignore = ["ANN001","ANN002", "ANN003", "ANN204", "ANN201"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]                 # no annotation requirement in test functions
"**/migrations/**" = ["ANN", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"


# ── djlint ───────────────────────────────────────────────────────────────────
[tool.djlint]
profile = "django"
indent = 2
max_line_length = 140
max_attribute_length = 100
format_css = true
format_js = true
ignore = "H006,H014,H025,H031,T003"
extension = "html"
